backend/requirements.txt
Django>=3.2,<4.0
djangorestframework
pycryptodome
qrcode[pil]
stripe>=2.55.0
